OFObject: Change type of -[hash] to unsigned long
The internal hash is still 32 bit in most places, but this way, it is at
least not baked into the API and ABI and can be upgraded later, should
that ever be necessary.

FossilOrigin-Name: aeb403a1ed86aa0e6461f35b1c3d21280f62dd054493ac0e96ae7e8af2dd5b11

+++ src/OFASN1Boolean.m
@@ -104,9 +104,9 @@
 	return true;
 }
 
-- (uint32_t)hash
+- (unsigned long)hash
 {
-	return (uint32_t)_booleanValue;
+	return _booleanValue;
 }
 
 - (OFString *)description
