+++ src/OFArray.m
@@ -515,7 +515,7 @@ static struct {
 	return true;
 }
 
-- (uint32_t)hash
+- (unsigned long)hash
 {
 	uint32_t hash;
